Add rocksdb dependencies
We tried to use the snarkos_storage crate that implements a wrapper of the rocksdb crate because of we may have to use it in the explorer project but we encountered a problem when saving information in the database.

In the snarkos_storage rocksdb wrapper the keys need to be u32 but in this toy example we need the keys to be String.

In order to solve this issue we now use regular rocksdb crate